requirements/base.txt
arrow==0.12.1
Babel==2.5.3
cmsplugin-filer==1.1.3
decorator==4.2.1
dj-database-url==0.5.0
Django==1.11.13 # pyup: <2.0
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.5.2
django-configurations==2.0
django-formtools==2.1
django-mptt==0.9.0
django-reversion==1.10.2 # pyup: >=1.8.2,<1.11
django-sekizai==0.10.0
djangocms-admin-style==1.2.8
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-link==2.1.2
django-parler==1.9.2
djangocms-picture==2.0.6
djangocms-snippet==2.0.0
djangocms-style==2.0.2
djangocms-text-ckeditor==3.6.0
djangocms-video==2.0.4
djangorestframework==3.8.2
easy-thumbnails==2.5
elasticsearch==6.2.0
gunicorn==19.8.1
lxml==4.2.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.4
raven==6.8.0
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
traitlets==4.3.2
tzlocal==1.5.1
urllib3==1.22
wcwidth==0.1.7
YURL==0.13
